StartPage copied to index.js and StartPage folder deleted
